check expand_as_v2 and less_equal
